# docs/swagger.yaml
swagger: "2.0"
info:
  title: Messaging System API
  version: 1.0.0
  description: Automatic message sending and tracking system API
  termsOfService: http://swagger.io/terms/
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: localhost:8080
basePath: /api
schemes:
  - http
  
paths:
  /service:
    post:
      summary: Controls the message sending service
      description: Starts or stops the message sending service
      tags:
        - service
      parameters:
        - name: action
          in: query
          required: true
          type: string
          description: "Service action: start or stop"
      responses:
        200:
          description: Successful response
          schema:
            type: object
            properties:
              success:
                type: boolean
              message:
                type: string
              running:
                type: boolean
        400:
          description: Invalid request
          schema:
            type: object
            properties:
              success:
                type: boolean
              error:
                type: string
        500:
          description: Server error
          schema:
            type: object
            properties:
              success:
                type: boolean
              error:
                type: string
  
  /service/status:
    get:
      summary: Gets the service status
      description: Retrieves the current status of the message service
      tags:
        - service
      responses:
        200:
          description: Successful response
          schema:
            type: object
            properties:
              success:
                type: boolean
              running:
                type: boolean
        500:
          description: Server error
          schema:
            type: object
            properties:
              success:
                type: boolean
              error:
                type: string
  
  /messages:
    get:
      summary: Lists sent messages
      description: Retrieves sent messages from the database with pagination support
      tags:
        - messages
      parameters:
        - name: page
          in: query
          required: false
          type: integer
          description: "Page number (default: 1)"
        - name: limit
          in: query
          required: false
          type: integer
          description: "Messages per page (default: 10)"
      responses:
        200:
          description: Successful response
          schema:
            type: object
            properties:
              success:
                type: boolean
              messages:
                type: array
                items:
                  $ref: '#/definitions/Message'
              total:
                type: integer
              page:
                type: integer
              limit:
                type: integer
              pages:
                type: integer
        500:
          description: Server error
          schema:
            type: object
            properties:
              success:
                type: boolean
              error:
                type: string

definitions:
  Message:
    type: object
    properties:
      id:
        type: integer
        format: int64
      content:
        type: string
      phoneNumber:
        type: string
      isSent:
        type: boolean
      sentAt:
        type: string
        format: date-time
      externalMsgId:
        type: string
      createdAt:
        type: string
        format: date-time
      updatedAt:
        type: string
        format: date-time 